generator/requirements.txt
declxml==1.1.3
names==0.3.0
psycopg2-binary==2.8.6

